Moved xmltramp.XmlTrampTests to a new module.
Code is not modified, only added two import lines.
